requirements.txt
fastmcp==0.1.0
python-dotenv==1.0.0
pathlib==1.0.1
typing-extensions==4.9.0
pydantic>=2.8.0
anyio>=4.5.0
click==8.1.7
colorama==0.4.6
idna==3.6
sniffio==1.3.1 